+++ packages/renderer/src/components/WatchPage/EpisodesList.vue
@@ -22,8 +22,8 @@
 import type {PropType} from 'vue';
 import {computed, defineComponent} from 'vue';
 import {useRoute} from 'vue-router';
-import type {Episode} from '/@/utils/videoProvider';
 import WinIcon from '/@/components/WinIcon.vue';
+import type {Episode} from '/@/utils/videoProvider';
 
 export default defineComponent({
   name: 'EpisodesList',
