+++ packages/renderer/src/utils/videoProvider.ts
@@ -1,10 +1,41 @@
 import {
   getEpisodes as providerGetEpisodes,
   getSeries as providerGetSeries,
-  getStream,
+  getStream as providerGetStream,
   getTranslations as providerGetTranslations,
 } from '/@/utils/providers/anime365';
-import type {Episode, Series, Translation, Video} from '/@/utils/ProviderInterfaces';
+
+interface HasID {
+  id: NumberLike
+}
+
+interface HasTitle {
+  title: string
+}
+
+export interface Series extends HasID, HasTitle {
+}
+
+export interface Episode extends HasID, HasTitle {
+  /**
+   * Порядковый номер эпизода в сериале
+   */
+  number: number
+}
+
+export type TranslationType = 'sub' | 'voice'
+
+export interface Translation extends HasID, HasTitle {
+  /**
+   * Тип перевода: Озвучка или субтитры
+   */
+  type: TranslationType
+}
+
+export interface Video {
+  quality: number
+  url: string
+}
 
 export function getSeries(id: NumberLike): Promise<Series | undefined> {
   return providerGetSeries(id);
@@ -21,5 +52,5 @@ export function getTranslations(providerEpisodeId: NumberLike): Promise<Translat
 
 
 export function getVideos(providerTranslationId: NumberLike): Promise<Video[]> {
-  return getStream(providerTranslationId).then(s => ([s]));
+  return providerGetStream(providerTranslationId).then(s => ([s]));
 }
